Update SocialMedia.sol
Modified onlyCommentOwner Modifer to the Check for Comment ID, and Post ID to Check if it is within the bounds of comments, and posts respectively. This prevents the contract from reverting due to attempting to access non-existent comments.

Also, add the Check which ensures only the owner of the comment can Modify it.